pyproject.toml
[project]
name = "rakuten-nint-analyzer"
version = "0.1.0"
description = "楽天市場・Nint連携型番商品リサーチ自動化ツール"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "playwright>=1.40.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "pillow>=10.1.0",
    "opencv-python>=4.8.0",
    "pandas>=2.1.0",
    "openpyxl>=3.1.0",
    "sqlalchemy>=2.0.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.0",
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.9"
strict = true
warn_return_any = true
warn_unused_configs = true
